[NFC] Move passes only used by GlobalOptimization out of Flow folders.
They are only used by GlobalOptimization pipeline, but not Flow
pipeline.
