+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/BUILD.bazel
@@ -32,7 +32,6 @@ iree_compiler_cc_library(
     srcs = [
         "AnnotateDispatches.cpp",
         "CaptureDispatchDynamicDims.cpp",
-        "CleanupNumericNarrowing.cpp",
         "CleanupTensorShapes.cpp",
         "CloneProducersIntoDispatchRegions.cpp",
         "CollapseDimensions.cpp",
@@ -42,21 +41,17 @@ iree_compiler_cc_library(
         "DeduplicateExecutables.cpp",
         "DispatchWithTransformDialect.cpp",
         "DumpDispatchGraph.cpp",
-        "ExpandTensorShapes.cpp",
         "ExportBenchmarkFuncs.cpp",
         "FoldUnitExtentDims.cpp",
         "FormDispatchRegions.cpp",
         "FormDispatchWorkgroups.cpp",
         "FormScalarDispatches.cpp",
         "FusionOfTensorOps.cpp",
-        "GeneralizeLinalgNamedOps.cpp",
-        "InferNumericNarrowing.cpp",
         "InitializeEmptyTensors.cpp",
         "InjectDispatchTracing.cpp",
         "InsertDispatchDebugTargets.cpp",
         "InterchangeGenericOps.cpp",
         "InterchangeTransposeGenericOps.cpp",
-        "OptimizeNumerics.cpp",
         "OutlineDispatchRegions.cpp",
         "PassDetail.h",
         "Passes.cpp",
