+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/CMakeLists.txt
@@ -31,7 +31,6 @@ iree_cc_library(
   SRCS
     "AnnotateDispatches.cpp"
     "CaptureDispatchDynamicDims.cpp"
-    "CleanupNumericNarrowing.cpp"
     "CleanupTensorShapes.cpp"
     "CloneProducersIntoDispatchRegions.cpp"
     "CollapseDimensions.cpp"
@@ -41,21 +40,17 @@ iree_cc_library(
     "DeduplicateExecutables.cpp"
     "DispatchWithTransformDialect.cpp"
     "DumpDispatchGraph.cpp"
-    "ExpandTensorShapes.cpp"
     "ExportBenchmarkFuncs.cpp"
     "FoldUnitExtentDims.cpp"
     "FormDispatchRegions.cpp"
     "FormDispatchWorkgroups.cpp"
     "FormScalarDispatches.cpp"
     "FusionOfTensorOps.cpp"
-    "GeneralizeLinalgNamedOps.cpp"
-    "InferNumericNarrowing.cpp"
     "InitializeEmptyTensors.cpp"
     "InjectDispatchTracing.cpp"
     "InsertDispatchDebugTargets.cpp"
     "InterchangeGenericOps.cpp"
     "InterchangeTransposeGenericOps.cpp"
-    "OptimizeNumerics.cpp"
     "OutlineDispatchRegions.cpp"
     "PassDetail.h"
     "Passes.cpp"
