+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/Passes.td
@@ -27,12 +27,6 @@ def CleanupTensorShapes :
   let constructor = "mlir::iree_compiler::IREE::Flow::createCleanupTensorShapesPass()";
 }
 
-def CleanupNumericNarrowing :
-    Pass<"iree-flow-cleanup-numeric-narrowing", ""> {
-  let summary = "Cleans up any numeric narrowing ops inserted by iree-flow-infer-numeric-narrowing";
-  let constructor = "mlir::iree_compiler::IREE::Flow::createCleanupNumericNarrowingPass()";
-}
-
 def CloneProducersIntoDispatchRegions :
     InterfacePass<"iree-flow-clone-producers-into-dispatch-regions", "mlir::FunctionOpInterface"> {
   let summary = "Clone producers into dispatch regions to be isolated above";
@@ -161,12 +155,6 @@ def DumpDispatchGraph : Pass<"iree-flow-dump-dispatch-graph-pass"> {
   let constructor = "mlir::iree_compiler::IREE::Flow::createDumpDispatchGraphPass()";
 }
 
-def ExpandTensorShapes :
-    Pass<"iree-flow-expand-tensor-shapes", "mlir::ModuleOp"> {
-  let summary = "Expands tensor shape dimensions into SSA values across the program.";
-  let constructor = "mlir::iree_compiler::IREE::Flow::createExpandTensorShapesPass()";
-}
-
 def ExportBenchmarkFuncs :
     Pass<"iree-flow-export-benchmark-funcs-pass", "mlir::ModuleOp"> {
   let summary = "Exports benchmark functions";
@@ -183,18 +171,6 @@ def FusionOfTensorOps :
     Option<"multiUseFusionIteration", "multi-use-fusion-iteration", "unsigned",
            /*default=*/"2", "Number of iterations to fuse multiuse ops">
   ];
-}
-
-def GeneralizeLinalgNamedOps :
-    InterfacePass<"iree-flow-generalize-linalg-named-ops", "mlir::FunctionOpInterface"> {
-  let summary = "Convert some Linalg named ops into linalg.generics";
-  let constructor = "mlir::iree_compiler::IREE::Flow::createGeneralizeLinalgNamedOpsPass()";
-}
-
-def InferNumericNarrowing :
-    Pass<"iree-flow-infer-numeric-narrowing", ""> {
-  let summary = "Infers and inserts util.numeric.optional_narrow ops at points that may be beneficial";
-  let constructor = "mlir::iree_compiler::IREE::Flow::createInferNumericNarrowingPass()";
 }
 
 def InitializeEmptyTensors :
@@ -251,12 +227,6 @@ def InterchangeTransposeGenericOps :
     Pass<"iree-flow-interchange-transpose-generic-ops", ""> {
   let summary = "Interchange transpose generic op loops to make the input indeximg map indentity.";
   let constructor = "mlir::iree_compiler::IREE::Flow::createInterchangeTransposeGenericOpsPass()";
-}
-
-def OptimizeNumerics :
-    Pass<"iree-flow-optimize-numerics", ""> {
-  let summary = "Optimizes numerics given annotations added via iree-flow-infer-numeric-narrowing";
-  let constructor = "mlir::iree_compiler::IREE::Flow::createOptimizeNumericsPass()";
 }
 
 def OutlineDispatchRegions :
