+++ compiler/src/iree/compiler/Dialect/Flow/Transforms/test/BUILD.bazel
@@ -18,7 +18,6 @@ iree_lit_test_suite(
         [
             "annotate_dispatches.mlir",
             "capture_dispatch_dynamic_dims.mlir",
-            "cleanup_numeric_narrowing.mlir",
             "cleanup_tensor_shapes.mlir",
             "clone_producers_into_dispatch_regions.mlir",
             "collapse_reduction.mlir",
@@ -29,21 +28,17 @@ iree_lit_test_suite(
             "dispatch_linalg_on_tensors_default.mlir",
             "dispatch_linalg_on_tensors_fusion_with_transpose.mlir",
             "dispatch_linalg_transform_dialect.mlir",
-            "expand_tensor_shapes.mlir",
             "export_benchmark_funcs.mlir",
             "fold_unit_dims.mlir",
             "form_dispatch_regions.mlir",
             "form_dispatch_workgroups.mlir",
             "form_scalar_dispatches.mlir",
             "fusion_of_tensor_ops.mlir",
-            "generalize_named_ops.mlir",
-            "infer_numeric_narrowing.mlir",
             "initialize_empty_tensors.mlir",
             "inject_dispatch_tracing.mlir",
             "insert_dispatch_debug_markers.mlir",
             "interchange_generic_ops.mlir",
             "interchange_transpose_generic_ops.mlir",
-            "optimize_numerics.mlir",
             "outline_dispatch_regions.mlir",
             "pad_fusion_with_consumer.mlir",
             "pad_fusion_with_producer.mlir",
