+++ compiler/src/iree/compiler/GlobalOptimization/CMakeLists.txt
@@ -40,13 +40,18 @@ iree_cc_library(
     "Passes.h"
     "Utils.h"
   SRCS
+    "CleanupNumericNarrowing.cpp"
     "Convert1X1FilterConv2DToMatmul.cpp"
     "DetachElementwiseFromNamedOps.cpp"
     "EraseUnusedLinalgOperands.cpp"
+    "ExpandTensorShapes.cpp"
     "ExpandVectors.cpp"
     "FuseDequantizationMatmul.cpp"
+    "GeneralizeLinalgNamedOps.cpp"
+    "InferNumericNarrowing.cpp"
     "LiftGenericToTransposeBatchMatmul.cpp"
     "MaterializeHomogeneousEncodings.cpp"
+    "OptimizeNumerics.cpp"
     "Passes.cpp"
     "RemoveZeroExtentTensors.cpp"
     "SetEncoding.cpp"
