+++ compiler/src/iree/compiler/GlobalOptimization/Passes.h
@@ -42,6 +42,10 @@ void buildGlobalOptimizationPassPipeline(
 // Input canonicalization and legalization
 //===----------------------------------------------------------------------===//
 
+// Cleans up any numeric narrowing ops inserted by
+// iree-global-opt-infer-numeric-narrowing.
+std::unique_ptr<Pass> createCleanupNumericNarrowingPass();
+
 // Creates a pass to convert linalg convolution ops with 1x1 kernels into
 // linalg.matmul
 std::unique_ptr<Pass> createConvert1X1FilterConv2DToMatmulPass();
@@ -54,6 +58,9 @@ std::unique_ptr<Pass> createDetachElementwiseFromNamedOpsPass();
 std::unique_ptr<OperationPass<mlir::ModuleOp>>
 createEraseUnusedLinalgOperands();
 
+// Expands tensor shape dimensions into SSA values across the program.
+std::unique_ptr<OperationPass<mlir::ModuleOp>> createExpandTensorShapesPass();
+
 // Expands vectors in vector/matrix operations into linalg.batch_matmul/matmul
 // forms.
 std::unique_ptr<Pass> createExpandVectorsPass();
@@ -63,10 +70,23 @@ std::unique_ptr<InterfacePass<mlir::FunctionOpInterface>>
 createFuseDequantizationMatmulPass(
     bool enableQuantizedMatmulReassociation = false);
 
+// Create a pass that generalizes some named Linalg ops into `linalg.generic`
+// operations since the IREE compiler can handle that better.
+std::unique_ptr<InterfacePass<mlir::FunctionOpInterface>>
+createGeneralizeLinalgNamedOpsPass();
+
+// Infers and inserts util.numeric.optional_narrow ops at points that may be
+// beneficial.
+std::unique_ptr<Pass> createInferNumericNarrowingPass();
+
 // Materializes logical encodings to physical encodings if there is a single
 // device target.
 std::unique_ptr<OperationPass<mlir::ModuleOp>>
 createMaterializeHomogeneousEncodingsPass();
+
+// Optimizes numerics given annotations added via
+// iree-global-opt-infer-numeric-narrowing.
+std::unique_ptr<Pass> createOptimizeNumericsPass();
 
 // Removes tensors that have 0-extents.
 std::unique_ptr<InterfacePass<mlir::FunctionOpInterface>>
