+++ compiler/src/iree/compiler/GlobalOptimization/test/BUILD.bazel
@@ -16,13 +16,18 @@ iree_lit_test_suite(
     name = "lit",
     srcs = enforce_glob(
         [
+            "cleanup_numeric_narrowing.mlir",
             "conv1x1_to_matmul.mlir",
             "detach_elementwise_from_named_ops.mlir",
+            "expand_tensor_shapes.mlir",
             "expand_vectors.mlir",
             "flow_hoist_into_globals.mlir",
             "fuse_dequantization_matmul.mlir",
+            "generalize_named_ops.mlir",
+            "infer_numeric_narrowing.mlir",
             "lift_generic_to_transpose_batch_matmul.mlir",
             "materialize_homogeneous_encodings.mlir",
+            "optimize_numerics.mlir",
             "remove_zero_extent_tensors.mlir",
             "set_encoding.mlir",
             "transformation_pipeline.mlir",
